- added text_input

+++ streamlit_app.py
@@ -15,13 +15,13 @@ yearly_oil = st.text_input('how many gallons of oil do you use per year for heat
 yearly_oil = str(yearly_oil)
 st.subheader('Transport:')
 total_yearly_mileage = st.number_input('how many miles have you done in your vehicle this year?: ', 100, 100000)
-total_yearly_mileage = int(total_yearly_mileage)
+total_yearly_mileage = float(total_yearly_mileage)
 total_yearly_gallons = total_yearly_mileage / 25
-total_yearly_gallons = int(total_yearly_gallons)
+total_yearly_gallons = float(total_yearly_gallons)
 number_of_flights_less = st.number_input('How many flights have you taken this year that are 4 hours or less?: ', 0, 50)
-number_of_flights_less = int(number_of_flights_less)
+number_of_flights_less = float(number_of_flights_less)
 number_of_flights_more = st.number_input('How many flights have you taken this year that are 4 hours or more?: ', 0, 50)
-number_of_flights_more = int(number_of_flights_more)
+number_of_flights_more = float(number_of_flights_more)
 st.subheader('Recycling')
 recycle_newspaper = st.selectbox('Do you recycle newspaper? (y/n): ', ['Yes', 'No'])
 recycle_newspaper = recycle_newspaper.lower()
@@ -29,7 +29,7 @@ recycle_aluminium = st.selectbox('Do you recycle aluminium and tin? (y/n): ', ['
 recycle_aluminium = recycle_aluminium.lower()
 
 trees = st.number_input("how many trees have you planted this year?: ", 0, 1000)
-trees = int(trees)
+trees = float(trees)
 
 carbon_offset = 46.2971
